+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 - Aliased dependency names in `package-lock.json`
+- Aliased dependency names in `yarn.lock`
 
 ## [5.8.1] - 2023-11-07